requirements.txt
discord==2.2.3
langchain==0.0.176
openai==0.27.7
python-dotenv==1.0.0

